backend/requirements.txt
aiofiles==23.1.0
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
async-timeout==4.0.2
attrs==23.1.0
backoff==2.2.1
beir==1.0.1
blis==0.7.9
catalogue==2.0.8
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
confection==0.0.4
cymem==2.0.7
dataclasses-json==0.5.7
datasets==2.12.0
defusedxml==0.7.1
dill==0.3.6
dnspython==2.3.0
easy-elasticsearch==0.0.9
elasticsearch==7.9.1
faiss-cpu==1.7.4
fastapi==0.95.1
filelock==3.12.0
frozenlist==1.3.3
fsspec==2023.5.0
fuzzywuzzy==0.18.0
gpl==0.1.4
gql==3.4.1
graphql-core==3.2.3
greenlet==2.0.2
h11==0.14.0
httpie==3.2.1
huggingface-hub==0.14.1
idna==3.4
iniconfig==2.0.0
Jinja2==3.1.2
joblib==1.2.0
langchain==0.0.170
langcodes==3.3.0
Levenshtein==0.21.0
llama-index==0.6.8
markdown-it-py==2.2.0
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
mdurl==0.1.2
motor==3.1.2
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.14
murmurhash==1.0.9
mypy-extensions==1.0.0
networkx==3.1
nltk==3.8.1
numexpr==2.8.4
numpy==1.24.2
openai==0.27.5
openapi-schema-pydantic==1.2.4
packaging==23.1
pandas==2.0.0
pathy==0.10.1
Pillow==9.5.0
pluggy==1.0.0
preshed==3.0.8
protobuf==4.23.0
pyarrow==12.0.0
pydantic==1.10.7
Pygments==2.15.1
pymongo==4.3.3
PySocks==1.7.1
pytest==7.3.1
python-dateutil==2.8.2
python-decouple==3.8
python-Levenshtein==0.21.0
python-multipart==0.0.6
pytrec-eval==0.5
pytz==2023.3
PyYAML==6.0
rapidfuzz==3.0.0
regex==2023.5.5
requests==2.28.2
requests-toolbelt==0.10.1
responses==0.18.0
rich==13.3.4
scikit-learn==1.2.2
scipy==1.10.1
sentence-transformers==2.2.2
sentencepiece==0.1.99
six==1.16.0
smart-open==6.3.0
sniffio==1.3.0
spacy==3.5.3
spacy-legacy==3.0.12
spacy-loggers==1.0.4
SQLAlchemy==2.0.13
srsly==2.4.6
starlette==0.26.1
sympy==1.12
tenacity==8.2.2
thinc==8.1.10
threadpoolctl==3.1.0
tiktoken==0.4.0
tokenizers==0.13.3
torch==2.0.1
torchvision==0.15.2
tqdm==4.65.0
transformers==4.29.2
typer==0.7.0
typing-inspect==0.8.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
uvicorn==0.21.1
wasabi==1.1.1
xxhash==3.2.0
yarl==1.9.2
